feat: faster fuzz way
