+++ Cargo.toml
@@ -17,7 +17,6 @@ members = [
     "libs/jwst-codec/fuzz",
     "libs/jwst-logger",
     "libs/jwst-rpc",
-    "libs/jwst-rpc/fuzz",
     "libs/jwst-storage",
     "libs/jwst-storage/src/migration",
 ]
