+++ libs/jwst-codec/Cargo.toml
@@ -15,6 +15,7 @@ nom = "7.1.3"
 rand = "0.8.5"
 serde_json = "1.0.94"
 thiserror = "1.0.40"
+y-sync = "0.3.0"
 
 # ======= workspace dependencies =======
 jwst-logger = { path = "../jwst-logger" }
@@ -22,8 +23,11 @@ jwst-logger = { path = "../jwst-logger" }
 [dev-dependencies]
 criterion = { version = "0.4.0", features = ["html_reports"] }
 lib0 = "0.16.5"
-serde = { version = "1.0.155", features = ["derive"] }
+proptest = "1.1.0"
+proptest-derive = "0.3.0"
 rand = "0.8.5"
+serde = { version = "1.0.155", features = ["derive"] }
+yrs = "0.16.5"
 
 [[bench]]
 name = "codec_benchmarks"
