+++ libs/jwst-codec/fuzz/Cargo.toml
@@ -21,6 +21,12 @@ test = false
 doc = false
 
 [[bin]]
+name = "sync_message"
+path = "fuzz_targets/sync_message.rs"
+test = false
+doc = false
+
+[[bin]]
 name = "i64_decode"
 path = "fuzz_targets/i64_decode.rs"
 test = false
