+++ libs/jwst-codec/src/lib.rs
@@ -1,11 +1,16 @@
 mod codec;
 mod doc;
+mod protocol;
 
 pub use codec::{
     read_var_buffer, read_var_i64, read_var_string, read_var_u64, write_var_buffer, write_var_i64,
     write_var_string, write_var_u64,
 };
 pub use doc::{read_update, Content, Doc, Id, Item, Update};
+pub use protocol::{
+    convert_awareness_update, read_sync_message, write_sync_message, AwarenessState,
+    AwarenessStates, SyncMessage,
+};
 
 use jwst_logger::warn;
 use nanoid::nanoid;
