+++ libs/jwst-codec/src/protocol/awareness.rs
