+++ libs/jwst-codec/src/protocol/doc.rs
