+++ libs/jwst-codec/src/protocol/mod.rs
@@ -7,14 +7,12 @@ mod utils;
 pub use awareness::{AwarenessState, AwarenessStates};
 pub use scanner::SyncMessageScanner;
 pub use sync::{read_sync_message, write_sync_message, SyncMessage};
+pub use utils::convert_awareness_update;
 
 use super::*;
 use awareness::{read_awareness, write_awareness};
 use doc::{read_doc_message, write_doc_message, DocMessage};
-use jwst_codec::{
-    read_var_buffer, read_var_string, read_var_u64, write_var_buffer, write_var_string,
-    write_var_u64,
-};
+use jwst_logger::debug;
 use nom::{
     error::{Error, ErrorKind},
     IResult,
