+++ libs/jwst-codec/src/protocol/scanner.rs
