--- libs/jwst-rpc/fuzz/.gitignore
@@ -1,4 +1,0 @@
-target
-corpus
-artifacts
-coverage
