--- libs/jwst-rpc/fuzz/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "jwst-rpc-fuzz"
-version = "0.0.0"
-publish = false
-edition = "2021"
-
-[package.metadata]
-cargo-fuzz = true
-
-[dependencies]
-libfuzzer-sys = "0.4"
-
-[dependencies.jwst-rpc]
-path = ".."
-default-features = false
-
-[profile.release]
-debug = 1
-
-[[bin]]
-name = "sync_message"
-path = "fuzz_targets/sync_message.rs"
-test = false
-doc = false
