+++ libs/jwst-rpc/src/broadcast.rs
@@ -1,5 +1,6 @@
 use super::*;
 use jwst::{sync_encode_update, Workspace};
+use jwst_codec::{convert_awareness_update, write_sync_message};
 use lru_time_cache::LruCache;
 use std::{collections::HashMap, sync::Mutex};
 use tokio::sync::{broadcast::Sender, RwLock};
