+++ libs/jwst-rpc/src/lib.rs
@@ -4,7 +4,6 @@ mod client;
 mod connector;
 mod context;
 mod handler;
-mod protocol;
 mod types;
 mod utils;
 
@@ -17,9 +16,6 @@ pub use broadcast::{BroadcastChannels, BroadcastType};
 pub use connector::memory_connector;
 pub use context::RpcContextImpl;
 pub use handler::handle_connector;
-pub use protocol::{
-    read_sync_message, write_sync_message, AwarenessState, AwarenessStates, SyncMessage,
-};
 pub use utils::{connect_memory_workspace, MinimumServerContext};
 
 use jwst::{debug, error, info, trace};
@@ -28,7 +24,6 @@ use tokio::{
     sync::mpsc::{Receiver, Sender},
     time::{sleep, Duration},
 };
-use utils::convert_awareness_update;
 
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub enum SyncState {
